refactor: [FE] EventType 수정
- 파일 관리를 위해 Event 관련 타입들을 모두 eventTypes.ts 파일로 이동
- 이벤트키 관리를 위해 EventKeyType 정의
