--- frontend/src/common/types/eventDataType.ts
@@ -1,12 +1,0 @@
-import {EventType} from './eventType';
-
-interface EventDataType{
-    eventTypes: EventType[];
-    eventKey: string;
-    listeners: EventListener[];
-    bindObj: Object;
-}
-
-export {
-    EventDataType
-}
