--- frontend/src/common/types/eventTargetDataType.ts
@@ -1,8 +1,0 @@
-interface EventTargetDataType{
-    listener: EventListener;
-    bindObj: Object;
-}
-
-export {
-    EventTargetDataType
-}
