--- frontend/src/common/types/eventType.ts
@@ -1,11 +1,0 @@
-enum EventType {
-    click = 'click',
-    keyup = 'keyup'
-}
-
-const eventTypes = ['click', 'keyup'];
-
-export {
-    EventType,
-    eventTypes
-}
